% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{factorize}
\alias{factorize}
\title{Factorize}
\usage{
factorize(data)
}
\arguments{
\item{data}{A Dataframe}
}
\value{
A Dataframe with all the character columns replaced by factor columns
}
\description{
Factorize
}

