% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{fixNAs}
\alias{fixNAs}
\title{FixNAs}
\usage{
fixNAs(data, replacement = "Unknown")
}
\arguments{
\item{data}{a Dataframe with some NAs to replace}
}
\value{
A dataframe without any NAs
}
\description{
Replaces missing values:
  Integer columns get the integer closest to the arithmetic mean of the column,
  Numeric columns get ther arithmetic mean of the column,
  Columns of nothing but NAs get deleted,
  And Character columns get the string 'Unknown'
}
