% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{partial_apply}
\alias{partial_apply}
\title{Partial Apply}
\usage{
partial_apply(FUN, ...)
}
\arguments{
\item{...}{}
}
\description{
Borrowed from the Functional Package, which had misnamed it 'Curry'
}
