% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{PointsToHexagons}
\alias{PointsToHexagons}
\title{Points to Hexagons}
\usage{
PointsToHexagons(points)
}
\arguments{
\item{points}{}
}
\description{
Takes a SpatialPoints object, draws hexagons around the points, and returns the SpatialPolygons object.
}
