% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{PointsToSquares}
\alias{PointsToSquares}
\title{Points to Squares}
\usage{
PointsToSquares(points)
}
\arguments{
\item{points}{}
}
\description{
Takes a SpatialPoints object, draws squares around the points, and returns the SpatialPolygons object.
}

