% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_file}
\alias{read_file}
\title{Read File}
\usage{
read_file(path, ...)
}
\arguments{
\item{...}{}
}
\description{
This insane thing is designed to reduce the number of
dependencies while maintaining a workflow that 'just works'. The tradeoff is
that it may behave differently on systems with different sets of packages.

TODO: think through how to do this properly, if possible.
}
