% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{recenter}
\alias{max0}
\alias{max100}
\alias{mean0}
\alias{median0}
\alias{min0}
\alias{recenter}
\title{Recentering Functions}
\usage{
recenter(x, diff)

max0(x)

max100(x)

mean0(x, diff)

median0(x)

min0(x)
}
\arguments{
\item{x}{a numeric vector}

\item{diff}{The number should be removed from each value of x}
}
\description{
A family of functions to "recenter" a vector around a new value.
This changes the ratios between values of x, while preserving
the differences between the values.
}

