% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\alias{rescale01}
\alias{rescale100}
\alias{rescale11}
\title{Rescaling functions}
\usage{
rescale(x, minx, maxx)

rescale01(x)

rescale11(x)

rescale100(x)
}
\arguments{
\item{x}{numeric vector}

\item{minx}{The desired minimum for the output vector}

\item{maxx}{The desired maximum for the output vector}
}
\description{
A Family of functions for transform the values of a vector
such that it falls into a specified range, without modifying
the ratios between values.
}

