% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%\\\%}
\alias{\%U\%}
\alias{\%\\\%}
\alias{\%n\%}
\title{Inline Set Operators}
\usage{
x \%\\\% y

x \%U\% y

x \%n\% y
}
\arguments{
\item{a}{a set}

\item{b}{another set}
}
\description{
Inline Set Operators
}

