% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stfu.R
\name{stfu}
\alias{stfu}
\title{Shut the File Up}
\usage{
stfu(expr)
}
\arguments{
\item{expr}{}
}
\value{
whatever expr should return
}
\description{
Given an expression, stfu evaluates and returns the value of the expression, but supressesses output to the console.
}
\examples{
> stfu(print("Foo"))
[1] "Foo"
> stfu(cat("Foo"))
NULL
}

